-- add - return java 17

+++ src/main/java/org/nickas21/smart/usr/entity/UsrTcpWiFiBmsSummary.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.time.Instant;
 
-public record UsrTcpWiFiBmsSummary(Instant timestamp, int socPercent, String bmsErrors, String bmsSummary) {
+public record UsrTcpWiFiBmsSummary(Instant timestamp, double socPercent, String bmsErrors, String bmsSummary) {
     @JsonCreator
     public UsrTcpWiFiBmsSummary(@JsonProperty("timestamp") Instant timestamp,
-                                 @JsonProperty("socPercent") int socPercent,
+                                 @JsonProperty("socPercent") double socPercent,
                                  @JsonProperty("bmsErrors") String bmsErrors,
                                  @JsonProperty("bmsSummary") String bmsSummary) {
         this.timestamp = timestamp;
